requirements.txt
fastapi
vellox
google-genai
httpx